after like added

+++ backend/routes/post.js
@@ -4,31 +4,28 @@ import { io } from "../server.js";
 import { prisma } from "../prisma.js";
 import {
   authentication,
-  authorization,
 } from "../middlewares/authMiddleware.js";
-
 /**
  * @swagger
  * /api/posts:
  *   post:
  *     summary: Create a new Post
+ *     security:
+ *       - BearerAuth: []
  *     tags: [Posts]
- *     parameters:
- *       - in: body
- *         name: body
- *         required: true
- *         schema:
- *           type: object
- *           properties:
- *             content:
- *               type: string
- *               description: the content of the post
- *             userId:
- *               type: string
- *               description: User Id of the post author
+ *     requestBody:
+ *       required: true
+ *       content:
+ *         application/json:
+ *           schema:
+ *             type: object
+ *             properties:
+ *               content:
+ *                 type: string
+ *                 description: The content of the post
  *     responses:
  *       201:
- *         description: post created successfully
+ *         description: Post created successfully
  *       401:
  *         description: Unauthorized - Invalid or missing token
  *       403:
@@ -36,21 +33,22 @@ import {
  *       500:
  *         description: Server Error
  */
-router.post( "/",
-  async (req, res) => {
-    const { content, userId } = req.body;
-    const post = await prisma.post.create({
-      data: { content, authorId: parseInt(userId) },
-    });
-    io.emit("newPost", post);
-    res.json({ post, user: req.user });
-  }
-);
+router.post("/", authentication, async (req, res) => {
+  const { userId } = req.user;
+  const { content } = req.body;
+  const post = await prisma.post.create({
+    data: { content, authorId: parseInt(userId) },
+  });
+  io.emit("newPost", post);
+  res.json({ post, user: req.user });
+});
 /**
  * @swagger
  * /api/posts/{id}:
  *   put:
- *     summary: Update a post
+ *     summary: Update a post(only by post owner and admin)
+ *     security:
+ *      - BearerAuth: []
  *     tags: [Posts]
  *     parameters:
  *       - in: path
@@ -78,21 +76,40 @@ router.post( "/",
  *         description: Post not found
  *       500:
  *         description: Server Error
-*/router.put("/:id",  async (req, res) => {
+ */ router.put("/:id", authentication, async (req, res) => {
   try {
     const { id } = req.params;
+    const { content } = req.body;
+    const userId = req.user.userId; 
+    const userRole = req.user.role; 
 
-    const { content } = req.body;  
-    const post = await prisma.post.update({
+    const post = await prisma.post.findUnique({
       where: { id: parseInt(id) },
-      data: { 
+    });
+
+    if (!post) {
+      return res.status(404).json({ error: "Post not found" });
+    }
+
+    if (post.authorId !== userId && userRole !== "admin") {
+      return res
+        .status(403)
+        .json({
+          error: "Forbidden - You do not have permission to update this post",
+        });
+    }
+
+    const updatedPost = await prisma.post.update({
+      where: { id: parseInt(id) },
+      data: {
         content: content,
       },
     });
-    io.emit("updatePost", post);                                                                                                                                                                                                                                                                                                                                                                                                                                                                                                              
-    res.json(post);
+
+    io.emit("updatePost", updatedPost);
+    res.json(updatedPost);
   } catch (error) {
-    console.log(error)
+    console.log(error);
     res.status(500).json({ error: "Failed to update post" });
   }
 });
@@ -100,7 +117,9 @@ router.post( "/",
  * @swagger
  * /api/posts/{id}:
  *   delete:
- *     summary: Delete a post by ID
+ *     summary: Delete a post by ID(only by post owner and admin)
+ *     security:
+ *      - BearerAuth: []
  *     tags: [Posts]
  *     parameters:
  *       - in: path
@@ -118,8 +137,24 @@ router.post( "/",
  *         description: Server Error
  */
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", authentication, async (req, res) => {
   const { id } = req.params;
+  const { userId ,role} = req.user;
+  const post = await prisma.post.findUnique({
+    where: { id: parseInt(id) },
+  });
+
+  if (!post) {
+    return res.status(404).json({ error: "Post not found" });
+  }
+
+  if (post.authorId !== userId && role !== "admin") {
+    return res
+      .status(403)
+      .json({
+        error: "Forbidden - You do not have permission to delete this post",
+      });
+  }
   await prisma.post.delete({
     where: { id: parseInt(id) },
   });
@@ -130,8 +165,6 @@ router.delete("/:id", async (req, res) => {
  * /api/posts:
  *   get:
  *     summary: Get all posts
- *     security:
- *      - BearerAuth: []
  *     tags: [Posts]
  *     responses:
  *       200:
@@ -140,29 +173,26 @@ router.delete("/:id", async (req, res) => {
  *         description: Server Error
  *
  */
-router.get(
-  "/",authentication,
-  async (req, res) => {
-    try {
-      const posts = await prisma.post.findMany({
-        orderBy: {
-          createdAt: "asc",
-        },
-        include: {
-          author: {
-            select: {
-              username: true,
-            },
+router.get("/", async (req, res) => {
+  try {
+    const posts = await prisma.post.findMany({
+      orderBy: {
+        createdAt: "asc",
+      },
+      include: {
+        author: {
+          select: {
+            username: true,
           },
         },
-      });
+      },
+    });
 
-      res.json(posts);
-    } catch (err) {
-      res.status(500).json({ message: "Server Error" });
-    }
+    res.json(posts);
+  } catch (err) {
+    res.status(500).json({ message: "Server Error" });
   }
-);
+});
 
 /**
  * @swagger
@@ -184,47 +214,44 @@ router.get(
  *         description: Server Error
  *
  */
-router.get(
-  "/:id",
-  async (req, res) => {
-    const { id } = req.params;
-    try {
-      const posts = await prisma.post.findFirst({
-        where: {
-          id: parseInt(id),
+router.get("/:id", async (req, res) => {
+  const { id } = req.params;
+  try {
+    const posts = await prisma.post.findFirst({
+      where: {
+        id: parseInt(id),
+      },
+      include: {
+        author: {
+          select: {
+            username: true,
+            role: true,
+            email: true,
+          },
         },
-        include: {
-          author: {
-            select: {
-              username: true,
-              role: true,
-              email:true
+        comments: {
+          select: {
+            content: true,
+            author: {
+              select: {
+                username: true,
+                role: true,
+                email: true,
+              },
             },
           },
-          comments: {
-            select: {
-              content: true,
-              author: {
-                select: {
-                  username: true,
-                  role: true,
-                  email:true
-                }
-              }
-            },
-           
-          }
         },
-      });
+        likes:true
+      },
+    });
 
-      res.json(posts);
-    } catch (err) {
-      console.log(err)
+    res.json(posts);
+  } catch (err) {
+    console.log(err);
 
-      res.status(500).json({ message: "Server Error" });
-    }
+    res.status(500).json({ message: "Server Error" });
   }
-);
+});
 
 /**
  * @swagger
@@ -241,34 +268,30 @@ router.get(
  *         description: Server Error happened
  *
  */
-router.get(
-  "/mypost/posts",authentication,
-  async (req, res) => {
-    const { userId } = req.user;
-    console.log(userId,"this is my id")
-    try {
-      const posts = await prisma.post.findMany({
-        where: {
-          authorId: parseInt(userId),
-        },
-        include: {
-          author: {
-            select: {
-              username: true,
-              email:true
-            },
+router.get("/mypost/posts", authentication, async (req, res) => {
+  const { userId } = req.user;
+  console.log(userId, "this is my id");
+  try {
+    const posts = await prisma.post.findMany({
+      where: {
+        authorId: parseInt(userId),
+      },
+      include: {
+        author: {
+          select: {
+            username: true,
+            email: true,
           },
         },
-      });
+      },
+    });
 
-      res.json(posts);
-    } catch (err) {
-      console.log(err)
-      res.status(500).json({ message: "Server Error hello",err });
-    }
+    res.json(posts);
+  } catch (err) {
+    console.log(err);
+    res.status(500).json({ message: "Server Error hello", err });
   }
-);
-
+});
 
 /**
  * @swagger
@@ -276,8 +299,6 @@ router.get(
  *   get:
  *     summary: Get  posts by author id
  *     tags: [Posts]
- *     security:
- *      - BearerAuth: []
  *     parameters:
  *       - in: path
  *         name: authorId
@@ -292,31 +313,27 @@ router.get(
  *         description: Server Error
  *
  */
-router.get(
-  "/by-author/:authorId",authentication,
-  async (req, res) => {
-    const { authorId } = req.params;
-    console.log(authorId)
-    try {
-      const posts = await prisma.post.findMany({
-        where: {
-          authorId: parseInt(authorId),
-        },
-        include: {
-          author: {
-            select: {
-              username: true,
-              email:true
-            },
+router.get("/by-author/:authorId", async (req, res) => {
+  const { authorId } = req.params;
+  try {
+    const posts = await prisma.post.findMany({
+      where: {
+        authorId: parseInt(authorId),
+      },
+      include: {
+        author: {
+          select: {
+            username: true,
+            email: true,
           },
         },
-      });
+      },
+    });
 
-      res.json(posts);
-    } catch (err) {
-      res.status(500).json({ message: "Server Error" });
-    }
+    res.json(posts);
+  } catch (err) {
+    res.status(500).json({ message: "Server Error" });
   }
-);
+});
 
 export default router;
